+++ sdmx-dl-format-util/src/main/java/sdmxdl/format/time/BasicTimePeriod.java
@@ -1,0 +1,8 @@
+package sdmxdl.format.time;
+
+/**
+ * <pre>
+ * </pre>
+ */
+public interface BasicTimePeriod extends StandardTimePeriod {
+}