+++ sdmx-dl-format-util/src/test/java/sdmxdl/format/time/StandardReportingPeriodTest.java
@@ -1,14 +1,13 @@
 package sdmxdl.format.time;
 
 import org.junit.jupiter.api.Test;
-import sdmxdl.format.time.StandardReportingPeriod;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class StandardReportingPeriodTest {
 
-    final StandardReportingPeriod s = StandardReportingPeriod
+    static final StandardReportingPeriod S = StandardReportingPeriod
             .builder()
             .reportingYear(2010)
             .periodIndicator('S')
@@ -16,7 +15,7 @@ public class StandardReportingPeriodTest {
             .periodValueDigits(1)
             .build();
 
-    final StandardReportingPeriod m2 = StandardReportingPeriod
+    static final StandardReportingPeriod M2 = StandardReportingPeriod
             .builder()
             .reportingYear(2010)
             .periodIndicator('M')
@@ -24,7 +23,7 @@ public class StandardReportingPeriodTest {
             .periodValueDigits(2)
             .build();
 
-    final StandardReportingPeriod m1 = StandardReportingPeriod
+    static final StandardReportingPeriod M1 = StandardReportingPeriod
             .builder()
             .reportingYear(2010)
             .periodIndicator('M')
@@ -34,15 +33,15 @@ public class StandardReportingPeriodTest {
 
     @Test
     public void testToString() {
-        assertThat(s.toString()).isEqualTo("2010-S2");
-        assertThat(m2.toString()).isEqualTo("2010-M03");
+        assertThat(S.toString()).isEqualTo("2010-S2");
+        assertThat(M2.toString()).isEqualTo("2010-M03");
     }
 
     @Test
     public void testParse() {
-        assertThat(StandardReportingPeriod.parse("2010-S2")).isEqualTo(s);
-        assertThat(StandardReportingPeriod.parse("2010-M03")).isEqualTo(m2);
-        assertThat(StandardReportingPeriod.parse("2010-M3")).isEqualTo(m1);
+        assertThat(StandardReportingPeriod.parse("2010-S2")).isEqualTo(S);
+        assertThat(StandardReportingPeriod.parse("2010-M03")).isEqualTo(M2);
+        assertThat(StandardReportingPeriod.parse("2010-M3")).isEqualTo(M1);
 
         assertThatIllegalArgumentException()
                 .isThrownBy(() -> StandardReportingPeriod.parse("2010-M"));