+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/provider/connectors/drivers/InseeDriver.java
@@ -17,7 +17,6 @@
 package internal.sdmxdl.provider.connectors.drivers;
 
 import internal.sdmxdl.provider.connectors.ConnectorsRestClient;
-import internal.sdmxdl.provider.connectors.Connectors;
 import internal.sdmxdl.provider.connectors.HasDetailSupported;
 import it.bancaditalia.oss.sdmx.api.Codelist;
 import it.bancaditalia.oss.sdmx.api.DataFlowStructure;
@@ -27,11 +26,10 @@ import it.bancaditalia.oss.sdmx.client.RestSdmxClient;
 import it.bancaditalia.oss.sdmx.exceptions.SdmxException;
 import nbbrd.io.text.Parser;
 import nbbrd.service.ServiceProvider;
-import sdmxdl.format.time.TimeFormats;
-import sdmxdl.provider.SdmxFix;
 import sdmxdl.format.ObsParser;
+import sdmxdl.format.ObsTimeParser;
 import sdmxdl.format.time.StandardReportingFormat;
-import sdmxdl.format.time.ObsTimeParser;
+import sdmxdl.provider.SdmxFix;
 import sdmxdl.provider.web.RestConnector;
 import sdmxdl.provider.web.WebDriverSupport;
 import sdmxdl.web.SdmxWebSource;
@@ -43,6 +41,7 @@ import java.util.function.Supplier;
 import java.util.logging.Level;
 
 import static internal.sdmxdl.provider.connectors.ConnectorsRestClient.CONNECTORS_CONNECTION_PROPERTIES;
+import static sdmxdl.format.ObsTimeParser.IGNORE_ERROR;
 import static sdmxdl.provider.SdmxFix.Category.CONTENT;
 
 /**
@@ -125,7 +124,7 @@ public final class InseeDriver implements WebDriver {
 //                if (!Connectors.isNoResultMatchingQuery(ex)) {
 //                    throw ex;
 //                }
-                LOGGER.log(Level.WARNING, "Cannot retrieve codes for ''{0}''", codelist.getFullIdentifier());
+            LOGGER.log(Level.WARNING, "Cannot retrieve codes for ''{0}''", codelist.getFullIdentifier());
 //            }
         }
     }
@@ -138,8 +137,8 @@ public final class InseeDriver implements WebDriver {
             .build();
 
     private static final ObsTimeParser EXTENDED_TIME_PARSER =
-            TimeFormats.OBSERVATIONAL_TIME_PERIOD
-                    .orElse(ObsTimeParser.onStandardReporting(REPORTING_TWO_MONTH));
+            ObsTimeParser.getObservationalTimePeriod(IGNORE_ERROR)
+                    .orElse(ObsTimeParser.onStandardReportingFormat(REPORTING_TWO_MONTH, IGNORE_ERROR));
 
     private static final Supplier<ObsParser> OBS_FACTORY = () -> new ObsParser(EXTENDED_TIME_PARSER, Parser.onDouble());
 }